chore(virtctl): Move expand command to own file
